Update README.md
The Oxford Homes Website project is a comprehensive platform designed to revolutionize the home buying and selling experience. As a software engineer, I developed this dynamic and user-friendly website to help a brokerage capture leads and facilitate seamless communication between prospective buyers and realtor partners.

The website's primary objective is to streamline the process of finding and acquiring homes, while providing a visually appealing and intuitive interface for users. Through extensive collaboration with the brokerage, I gained insights into their specific requirements and challenges, which guided the development process.